pyproject.toml
[tool.poetry]
name = "mybot"
version = "0.1.0"
description = ""
authors = ["xyrlsz"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
bilibili-api-python = "^16.1.1"
requests = "^2.31.0"
asyncio = "^3.4.3"
pyjson = "^1.4.1"
pyyaml = "^6.0.1"
websockets = "^12.0"
qrcode = "^7.4.2"
psutil = "^5.9.7"
pixivpy3 = "^3.7.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
